Prevent extra queries during prune revisions, potentially other instances
I noticed while debugging why the `utils/prune-revisions` console command was running so slow on a site with tens of millions of relations. After debugging, I realized disabling Scout caused the command to run blazing fast. Further digging made me realize that the getRelatedElements call was getting called with every single deleted element (and subelements, such as matrix or supertable elements) because there was no check to see if the element should be searchable to begin with. This commit adds in that check, causing the sync config item, propogating element state, and draft/revisionness to be respected during deletion where it was not before.